backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
motor==3.6.1
pymongo==4.9.2
pydantic==2.10.4
pydantic-settings==2.7.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
numpy==2.2.1
scikit-learn==1.6.0
requests==2.32.3
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.25.0
